+++ condimarket_front/lib/screens/login_screen.dart
@@ -1,0 +1,340 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/gestures.dart';
+import '../services/loginauth_service.dart';
+
+class LoginScreen extends StatefulWidget {
+  const LoginScreen({Key? key}) : super(key: key);
+
+  @override
+  _LoginScreenState createState() => _LoginScreenState();
+}
+
+class _LoginScreenState extends State<LoginScreen> {
+  // Método para mostrar información del usuario después del login exitoso
+  void _mostrarInformacionUsuario(Map<String, dynamic> userData) {
+    showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        title: Text('Inicio de sesión exitoso'),
+        content: SingleChildScrollView(
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              Text('¡Bienvenido, ${userData['nombre']}!'),
+              SizedBox(height: 8),
+              Text('ID: ${userData['id']}'),
+              Text('Email: ${userData['email']}'),
+              if (userData['telefono'] != null) Text('Teléfono: ${userData['telefono']}'),
+              if (userData['direccion'] != null) Text('Dirección: ${userData['direccion']}'),
+              Text('Rol: ${userData['rol']}'),
+            ],
+          ),
+        ),
+        actions: [
+          TextButton(
+            onPressed: () {
+              Navigator.of(context).pop();
+            },
+            child: Text('Aceptar'),
+          ),
+        ],
+      ),
+    );
+  }
+  final _formKey = GlobalKey<FormState>();
+  final _emailController = TextEditingController();
+  final _passwordController = TextEditingController();
+
+  bool _obscurePassword = true;
+  bool _isLoading = false;
+  String _errorMessage = '';
+
+  // Instancia del servicio de autenticación (cambiar a true para usar datos quemados)
+  final loginAuthService = LoginAuthService(usarDatosMock: true);
+  @override
+  void dispose() {
+    _emailController.dispose();
+    _passwordController.dispose();
+    super.dispose();
+  }
+
+  // Método para validar el formulario e iniciar sesión
+  Future<void> _login() async {
+    // Quitar el foco para cerrar el teclado
+    FocusScope.of(context).unfocus();
+
+    // Validar el formulario
+    if (!_formKey.currentState!.validate()) {
+      return;
+    }
+
+    setState(() {
+      _isLoading = true;
+      _errorMessage = '';
+    });
+
+    try {
+      final result = await loginAuthService.iniciarSesion(
+      email: _emailController.text.trim(),
+        password: _passwordController.text,
+      );
+
+      if (result['success']) {
+        // Inicio de sesión exitoso
+        // Mostrar información y enviar a la pantalla principal que debería estar implementada
+        // en otra parte de la aplicación
+        ScaffoldMessenger.of(context).showSnackBar(
+          SnackBar(
+            content: Text('Inicio de sesión exitoso'),
+            backgroundColor: Colors.green,
+          ),
+        );
+
+        // Aquí se debería navegar a la pantalla principal
+        // Por ahora, solo mostramos los datos del usuario
+        _mostrarInformacionUsuario(result['user']);
+      } else {
+        setState(() {
+          _errorMessage = result['message'];
+        });
+      }
+    } catch (e) {
+      setState(() {
+        _errorMessage = 'Error de conexión. Intente nuevamente.';
+      });
+    } finally {
+      setState(() {
+        _isLoading = false;
+      });
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Colors.white,
+      body: SafeArea(
+        child: Center(
+          child: SingleChildScrollView(
+            child: Padding(
+              padding: const EdgeInsets.all(24.0),
+              child: Form(
+                key: _formKey,
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  crossAxisAlignment: CrossAxisAlignment.stretch,
+                  children: [
+                    // Logo y título
+                    Image.asset(
+                      'assets/images/logo.png',
+                      height: 120,
+                      // Si no tienes el logo, puedes usar un Icon en su lugar
+                      errorBuilder: (context, error, stackTrace) => Icon(
+                        Icons.shopping_basket,
+                        size: 80,
+                        color: Colors.orange,
+                      ),
+                    ),
+                    const SizedBox(height: 24),
+                    Text(
+                      'CondiMarket',
+                      textAlign: TextAlign.center,
+                      style: TextStyle(
+                        fontSize: 28,
+                        fontWeight: FontWeight.bold,
+                        color: Colors.orange,
+                      ),
+                    ),
+                    const SizedBox(height: 12),
+                    Text(
+                      'Inicia sesión para continuar',
+                      textAlign: TextAlign.center,
+                      style: TextStyle(
+                        fontSize: 16,
+                        color: Colors.grey[600],
+                      ),
+                    ),
+                    const SizedBox(height: 36),
+
+                    // Formulario de inicio de sesión
+                    // Campo de email
+                    TextFormField(
+                      controller: _emailController,
+                      keyboardType: TextInputType.emailAddress,
+                      decoration: InputDecoration(
+                        labelText: 'Correo electrónico',
+                        prefixIcon: Icon(Icons.email_outlined),
+                        border: OutlineInputBorder(
+                          borderRadius: BorderRadius.circular(12),
+                        ),
+                        filled: true,
+                        fillColor: Colors.grey[50],
+                      ),
+                      validator: (value) {
+                        if (value == null || value.isEmpty) {
+                          return 'Por favor ingresa tu correo electrónico';
+                        }
+                        if (!RegExp(r'^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$').hasMatch(value)) {
+                          return 'Ingresa un correo electrónico válido';
+                        }
+                        return null;
+                      },
+                    ),
+                    const SizedBox(height: 16),
+
+                    // Campo de contraseña
+                    TextFormField(
+                      controller: _passwordController,
+                      obscureText: _obscurePassword,
+                      decoration: InputDecoration(
+                        labelText: 'Contraseña',
+                        prefixIcon: Icon(Icons.lock_outline),
+                        suffixIcon: IconButton(
+                          icon: Icon(
+                            _obscurePassword ? Icons.visibility_off : Icons.visibility,
+                          ),
+                          onPressed: () {
+                            setState(() {
+                              _obscurePassword = !_obscurePassword;
+                            });
+                          },
+                        ),
+                        border: OutlineInputBorder(
+                          borderRadius: BorderRadius.circular(12),
+                        ),
+                        filled: true,
+                        fillColor: Colors.grey[50],
+                      ),
+                      validator: (value) {
+                        if (value == null || value.isEmpty) {
+                          return 'Por favor ingresa tu contraseña';
+                        }
+                        return null;
+                      },
+                    ),
+                    const SizedBox(height: 12),
+
+                    // Link para olvidé mi contraseña
+                    Align(
+                      alignment: Alignment.centerRight,
+                      child: TextButton(
+                        onPressed: () {
+                          // Implementar recuperación de contraseña
+                          ScaffoldMessenger.of(context).showSnackBar(
+                            SnackBar(content: Text('Función próximamente disponible')),
+                          );
+                        },
+                        child: Text(
+                          'Olvidé mi contraseña',
+                          style: TextStyle(
+                            color: Colors.blue[700],
+                          ),
+                        ),
+                      ),
+                    ),
+                    const SizedBox(height: 12),
+
+                    // Mensaje de error si existe
+                    if (_errorMessage.isNotEmpty)
+                      Container(
+                        padding: EdgeInsets.all(12),
+                        decoration: BoxDecoration(
+                          color: Colors.red[50],
+                          borderRadius: BorderRadius.circular(8),
+                        ),
+                        child: Text(
+                          _errorMessage,
+                          style: TextStyle(
+                            color: Colors.red,
+                            fontSize: 14,
+                          ),
+                          textAlign: TextAlign.center,
+                        ),
+                      ),
+                    const SizedBox(height: 24),
+
+                    // Botón de inicio de sesión
+                    SizedBox(
+                      height: 54,
+                      child: ElevatedButton(
+                        onPressed: _isLoading ? null : _login,
+                        style: ElevatedButton.styleFrom(
+                          backgroundColor: Colors.orange,
+                          foregroundColor: Colors.white,
+                          shape: RoundedRectangleBorder(
+                            borderRadius: BorderRadius.circular(12),
+                          ),
+                          disabledBackgroundColor: Colors.orange.withOpacity(0.6),
+                        ),
+                        child: _isLoading
+                            ? CircularProgressIndicator(color: Colors.white)
+                            : Text(
+                          'Iniciar Sesión',
+                          style: TextStyle(
+                            fontSize: 16,
+                            fontWeight: FontWeight.bold,
+                          ),
+                        ),
+                      ),
+                    ),
+                    const SizedBox(height: 20),
+
+                    // Opción para registrarse
+                    RichText(
+                      textAlign: TextAlign.center,
+                      text: TextSpan(
+                        style: TextStyle(
+                          fontSize: 14,
+                          color: Colors.grey[700],
+                        ),
+                        children: [
+                          TextSpan(text: '¿No tienes cuenta? '),
+                          TextSpan(
+                            text: 'Regístrate',
+                            style: TextStyle(
+                              color: Colors.blue[700],
+                              fontWeight: FontWeight.bold,
+                            ),
+                            recognizer: TapGestureRecognizer()
+                              ..onTap = () {
+                                // Aquí iría la navegación a la pantalla de registro
+                                // pero como solo estamos implementando el inicio de sesión,
+                                // simplemente mostramos un mensaje
+                                ScaffoldMessenger.of(context).showSnackBar(
+                                  SnackBar(content: Text('Registro próximamente disponible')),
+                                );
+                              },
+                          ),
+                        ],
+                      ),
+                    ),
+
+                    // Opción de continuar como invitado
+                    TextButton(
+                      onPressed: () {
+                        // Aquí se debería navegar a la pantalla principal sin autenticación
+                        ScaffoldMessenger.of(context).showSnackBar(
+                          SnackBar(
+                            content: Text('Continuando como invitado'),
+                            backgroundColor: Colors.blueGrey,
+                          ),
+                        );
+                      },
+                      child: Text(
+                        'Continuar como invitado',
+                        style: TextStyle(
+                          color: Colors.grey[700],
+                        ),
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
