--- condimarket_front/lib/screens/register_screen.dart
@@ -1,457 +1,0 @@
-// register_screen.dart
-import 'package:flutter/material.dart';
-import 'package:flutter/gestures.dart';
-import '../services/auth_service.dart';
-
-class RegisterScreen extends StatefulWidget {
-  const RegisterScreen({Key? key}) : super(key: key);
-
-  @override
-  _RegisterScreenState createState() => _RegisterScreenState();
-}
-
-class _RegisterScreenState extends State<RegisterScreen> {
-  final _formKey = GlobalKey<FormState>();
-  final _nombreController = TextEditingController();
-  final _emailController = TextEditingController();
-  final _passwordController = TextEditingController();
-  final _confirmPasswordController = TextEditingController();
-  final _telefonoController = TextEditingController();
-  final _direccionController = TextEditingController();
-
-  bool _obscurePassword = true;
-  bool _obscureConfirmPassword = true;
-  bool _isLoading = false;
-  String _errorMessage = '';
-  bool _aceptaTerminos = false;
-
-  // Instancia del servicio de autenticación
-  final AuthService _authService = AuthService(usarDatosMock: true);
-
-  @override
-  void dispose() {
-    _nombreController.dispose();
-    _emailController.dispose();
-    _passwordController.dispose();
-    _confirmPasswordController.dispose();
-    _telefonoController.dispose();
-    _direccionController.dispose();
-    super.dispose();
-  }
-
-  // Método para mostrar información del usuario después del registro exitoso
-  void _mostrarInformacionUsuario(Map<String, dynamic> userData) {
-    showDialog(
-      context: context,
-      builder: (context) => AlertDialog(
-        title: Text('Registro exitoso'),
-        content: SingleChildScrollView(
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              Text('¡Bienvenido, ${userData['nombre']}!'),
-              SizedBox(height: 8),
-              Text('ID: ${userData['id']}'),
-              Text('Email: ${userData['email']}'),
-              if (userData['telefono'] != null) Text('Teléfono: ${userData['telefono']}'),
-              if (userData['direccion'] != null) Text('Dirección: ${userData['direccion']}'),
-              Text('Rol: ${userData['rol']}'),
-            ],
-          ),
-        ),
-        actions: [
-          TextButton(
-            onPressed: () {
-              // Cerrar el diálogo
-              Navigator.of(context).pop();
-            },
-            child: Text('Continuar'),
-          ),
-        ],
-      ),
-    );
-  }
-
-  // Método para validar el formulario y registrar al usuario
-  Future<void> _register() async {
-    // Quitar el foco para cerrar el teclado
-    FocusScope.of(context).unfocus();
-
-    // Validar el formulario
-    if (!_formKey.currentState!.validate()) {
-      return;
-    }
-
-    // Verificar que haya aceptado los términos y condiciones
-    if (!_aceptaTerminos) {
-      setState(() {
-        _errorMessage = 'Debes aceptar los términos y condiciones para continuar';
-      });
-      return;
-    }
-
-    setState(() {
-      _isLoading = true;
-      _errorMessage = '';
-    });
-
-    try {
-      final result = await _authService.registrarUsuario(
-        nombre: _nombreController.text.trim(),
-        email: _emailController.text.trim(),
-        password: _passwordController.text,
-        telefono: _telefonoController.text.trim(),
-        direccion: _direccionController.text.trim(),
-      );
-
-      if (result['success']) {
-        // Registro exitoso
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(
-            content: Text('Registro exitoso'),
-            backgroundColor: Colors.green,
-          ),
-        );
-
-        // Mostrar los datos del usuario registrado
-        _mostrarInformacionUsuario(result['user']);
-      } else {
-        setState(() {
-          _errorMessage = result['message'];
-        });
-      }
-    } catch (e) {
-      setState(() {
-        _errorMessage = 'Error de conexión. Intente nuevamente.';
-      });
-    } finally {
-      setState(() {
-        _isLoading = false;
-      });
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.white,
-      appBar: AppBar(
-        backgroundColor: Colors.white,
-        elevation: 0,
-        title: Text(
-          'Crear cuenta',
-          style: TextStyle(
-            color: Colors.black,
-            fontWeight: FontWeight.bold,
-          ),
-        ),
-      ),
-      body: SafeArea(
-        child: Center(
-          child: SingleChildScrollView(
-            child: Padding(
-              padding: const EdgeInsets.all(24.0),
-              child: Form(
-                key: _formKey,
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  crossAxisAlignment: CrossAxisAlignment.stretch,
-                  children: [
-                    // Logo
-                    Image.asset(
-                      'assets/images/logo.png',
-                      height: 100,
-                      errorBuilder: (context, error, stackTrace) => Icon(
-                        Icons.shopping_basket,
-                        size: 60,
-                        color: Colors.orange,
-                      ),
-                    ),
-                    const SizedBox(height: 16),
-                    Text(
-                      'CondiMarket',
-                      textAlign: TextAlign.center,
-                      style: TextStyle(
-                        fontSize: 24,
-                        fontWeight: FontWeight.bold,
-                        color: Colors.orange,
-                      ),
-                    ),
-                    const SizedBox(height: 10),
-                    Text(
-                      'Crea tu cuenta para disfrutar de todas las ventajas',
-                      textAlign: TextAlign.center,
-                      style: TextStyle(
-                        fontSize: 14,
-                        color: Colors.grey[600],
-                      ),
-                    ),
-                    const SizedBox(height: 24),
-
-                    // Formulario de registro
-                    // Campo de nombre
-                    TextFormField(
-                      controller: _nombreController,
-                      decoration: InputDecoration(
-                        labelText: 'Nombre completo',
-                        prefixIcon: Icon(Icons.person_outline),
-                        border: OutlineInputBorder(
-                          borderRadius: BorderRadius.circular(12),
-                        ),
-                        filled: true,
-                        fillColor: Colors.grey[50],
-                      ),
-                      validator: (value) {
-                        if (value == null || value.isEmpty) {
-                          return 'Por favor ingresa tu nombre';
-                        }
-                        return null;
-                      },
-                    ),
-                    const SizedBox(height: 16),
-
-                    // Campo de email
-                    TextFormField(
-                      controller: _emailController,
-                      keyboardType: TextInputType.emailAddress,
-                      decoration: InputDecoration(
-                        labelText: 'Correo electrónico',
-                        prefixIcon: Icon(Icons.email_outlined),
-                        border: OutlineInputBorder(
-                          borderRadius: BorderRadius.circular(12),
-                        ),
-                        filled: true,
-                        fillColor: Colors.grey[50],
-                      ),
-                      validator: (value) {
-                        if (value == null || value.isEmpty) {
-                          return 'Por favor ingresa tu correo electrónico';
-                        }
-                        if (!RegExp(r'^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$').hasMatch(value)) {
-                          return 'Ingresa un correo electrónico válido';
-                        }
-                        return null;
-                      },
-                    ),
-                    const SizedBox(height: 16),
-
-                    // Campo de teléfono (opcional)
-                    TextFormField(
-                      controller: _telefonoController,
-                      keyboardType: TextInputType.phone,
-                      decoration: InputDecoration(
-                        labelText: 'Teléfono (opcional)',
-                        prefixIcon: Icon(Icons.phone_outlined),
-                        border: OutlineInputBorder(
-                          borderRadius: BorderRadius.circular(12),
-                        ),
-                        filled: true,
-                        fillColor: Colors.grey[50],
-                      ),
-                      // No necesita validador porque es opcional
-                    ),
-                    const SizedBox(height: 16),
-
-                    // Campo de dirección (opcional)
-                    TextFormField(
-                      controller: _direccionController,
-                      keyboardType: TextInputType.streetAddress,
-                      decoration: InputDecoration(
-                        labelText: 'Dirección (opcional)',
-                        prefixIcon: Icon(Icons.home_outlined),
-                        border: OutlineInputBorder(
-                          borderRadius: BorderRadius.circular(12),
-                        ),
-                        filled: true,
-                        fillColor: Colors.grey[50],
-                      ),
-                      // No necesita validador porque es opcional
-                    ),
-                    const SizedBox(height: 16),
-
-                    // Campo de contraseña
-                    TextFormField(
-                      controller: _passwordController,
-                      obscureText: _obscurePassword,
-                      decoration: InputDecoration(
-                        labelText: 'Contraseña',
-                        prefixIcon: Icon(Icons.lock_outline),
-                        suffixIcon: IconButton(
-                          icon: Icon(
-                            _obscurePassword ? Icons.visibility_off : Icons.visibility,
-                          ),
-                          onPressed: () {
-                            setState(() {
-                              _obscurePassword = !_obscurePassword;
-                            });
-                          },
-                        ),
-                        border: OutlineInputBorder(
-                          borderRadius: BorderRadius.circular(12),
-                        ),
-                        filled: true,
-                        fillColor: Colors.grey[50],
-                      ),
-                      validator: (value) {
-                        if (value == null || value.isEmpty) {
-                          return 'Por favor ingresa una contraseña';
-                        }
-                        if (value.length < 6) {
-                          return 'La contraseña debe tener al menos 6 caracteres';
-                        }
-                        return null;
-                      },
-                    ),
-                    const SizedBox(height: 16),
-
-                    // Campo de confirmar contraseña
-                    TextFormField(
-                      controller: _confirmPasswordController,
-                      obscureText: _obscureConfirmPassword,
-                      decoration: InputDecoration(
-                        labelText: 'Confirmar contraseña',
-                        prefixIcon: Icon(Icons.lock_outline),
-                        suffixIcon: IconButton(
-                          icon: Icon(
-                            _obscureConfirmPassword ? Icons.visibility_off : Icons.visibility,
-                          ),
-                          onPressed: () {
-                            setState(() {
-                              _obscureConfirmPassword = !_obscureConfirmPassword;
-                            });
-                          },
-                        ),
-                        border: OutlineInputBorder(
-                          borderRadius: BorderRadius.circular(12),
-                        ),
-                        filled: true,
-                        fillColor: Colors.grey[50],
-                      ),
-                      validator: (value) {
-                        if (value == null || value.isEmpty) {
-                          return 'Por favor confirma tu contraseña';
-                        }
-                        if (value != _passwordController.text) {
-                          return 'Las contraseñas no coinciden';
-                        }
-                        return null;
-                      },
-                    ),
-                    const SizedBox(height: 16),
-
-                    // Checkbox para aceptar términos y condiciones
-                    Row(
-                      children: [
-                        Checkbox(
-                          value: _aceptaTerminos,
-                          onChanged: (value) {
-                            setState(() {
-                              _aceptaTerminos = value ?? false;
-                            });
-                          },
-                          activeColor: Colors.orange,
-                        ),
-                        Expanded(
-                          child: RichText(
-                            text: TextSpan(
-                              style: TextStyle(
-                                fontSize: 14,
-                                color: Colors.grey[700],
-                              ),
-                              children: [
-                                TextSpan(text: 'Acepto los '),
-                                TextSpan(
-                                  text: 'Términos y Condiciones',
-                                  style: TextStyle(
-                                    color: Colors.blue[700],
-                                    fontWeight: FontWeight.bold,
-                                  ),
-                                  recognizer: TapGestureRecognizer()
-                                    ..onTap = () {
-                                      // Mostrar términos y condiciones
-                                      showDialog(
-                                        context: context,
-                                        builder: (context) => AlertDialog(
-                                          title: Text('Términos y Condiciones'),
-                                          content: SingleChildScrollView(
-                                            child: Text(
-                                              'Estos son los términos y condiciones de CondiMarket. '
-                                                  'Al registrarte aceptas nuestras políticas de uso y privacidad. '
-                                                  'Tus datos serán utilizados únicamente para mejorar tu experiencia de compra.',
-                                              style: TextStyle(fontSize: 14),
-                                            ),
-                                          ),
-                                          actions: [
-                                            TextButton(
-                                              onPressed: () {
-                                                Navigator.of(context).pop();
-                                              },
-                                              child: Text('Aceptar'),
-                                            ),
-                                          ],
-                                        ),
-                                      );
-                                    },
-                                ),
-                              ],
-                            ),
-                          ),
-                        ),
-                      ],
-                    ),
-                    const SizedBox(height: 12),
-
-                    // Mensaje de error si existe
-                    if (_errorMessage.isNotEmpty)
-                      Container(
-                        padding: EdgeInsets.all(12),
-                        decoration: BoxDecoration(
-                          color: Colors.red[50],
-                          borderRadius: BorderRadius.circular(8),
-                        ),
-                        child: Text(
-                          _errorMessage,
-                          style: TextStyle(
-                            color: Colors.red,
-                            fontSize: 14,
-                          ),
-                          textAlign: TextAlign.center,
-                        ),
-                      ),
-                    const SizedBox(height: 24),
-
-                    // Botón de registro
-                    SizedBox(
-                      height: 54,
-                      child: ElevatedButton(
-                        onPressed: _isLoading ? null : _register,
-                        style: ElevatedButton.styleFrom(
-                          backgroundColor: Colors.orange,
-                          foregroundColor: Colors.white,
-                          shape: RoundedRectangleBorder(
-                            borderRadius: BorderRadius.circular(12),
-                          ),
-                          disabledBackgroundColor: Colors.orange.withOpacity(0.6),
-                        ),
-                        child: _isLoading
-                            ? CircularProgressIndicator(color: Colors.white)
-                            : Text(
-                          'Registrarme',
-                          style: TextStyle(
-                            fontSize: 16,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
